Bootstrap Density Tibble

Add density information to the output of tidy_bootstrap(), and bootstrap_unnest_tbl().

Usage

bootstrap_density_augment(.data)

Arguments

.data

The data that is passed from the tidy_bootstrap() or bootstrap_unnest_tbl() functions.

Value

A tibble

Details

This function takes as input the output of the tidy_bootstrap() or bootstrap_unnest_tbl() and returns an augmented tibble that has the following columns added to it: x, y, dx, and dy.

It looks for an attribute that comes from using tidy_bootstrap() or bootstrap_unnest_tbl() so it will not work unless the data comes from one of those functions.

See also

Author

Steven P. Sanderson II, MPH

Examples

x <- mtcars$mpg

tidy_bootstrap(x) %>%
  bootstrap_density_augment()
#> # A tibble: 50,000 × 5
#>    sim_number     x     y    dx       dy
#>    <fct>      <int> <dbl> <dbl>    <dbl>
#>  1 1              1  33.9  7.48 0.000194
#>  2 1              2  19.7  8.82 0.00127 
#>  3 1              3  18.1 10.2  0.00534 
#>  4 1              4  21   11.5  0.0152  
#>  5 1              5  32.4 12.9  0.0310  
#>  6 1              6  16.4 14.2  0.0500  
#>  7 1              7  17.3 15.5  0.0695  
#>  8 1              8  17.3 16.9  0.0861  
#>  9 1              9  32.4 18.2  0.0933  
#> 10 1             10  24.4 19.6  0.0855  
#> # … with 49,990 more rows

tidy_bootstrap(x) %>%
  bootstrap_unnest_tbl() %>%
  bootstrap_density_augment()
#> # A tibble: 50,000 × 5
#>    sim_number     x     y    dx       dy
#>    <fct>      <int> <dbl> <dbl>    <dbl>
#>  1 1              1  19.2  3.84 0.000329
#>  2 1              2  15.8  5.36 0.00208 
#>  3 1              3  21    6.89 0.00812 
#>  4 1              4  15.2  8.42 0.0198  
#>  5 1              5  27.3  9.94 0.0313  
#>  6 1              6  21.4 11.5  0.0371  
#>  7 1              7  22.8 13.0  0.0438  
#>  8 1              8  10.4 14.5  0.0562  
#>  9 1              9  19.2 16.0  0.0642  
#> 10 1             10  10.4 17.6  0.0632  
#> # … with 49,990 more rows