Unnest Tidy Bootstrap Tibble

Unnest the data output from tidy_bootstrap().

bootstrap_unnest_tbl(.data)

Arguments

.data

The data that is passed from the tidy_bootstrap() function.

Value

A tibble

Details

This function takes as input the output of the tidy_bootstrap() function and returns a two column tibble. The columns are sim_number and y

It looks for an attribute that comes from using tidy_bootstrap() so it will not work unless the data comes from that function.

See also

Author

Steven P. Sanderson II, MPH

Examples

tb <- tidy_bootstrap(.x = mtcars$mpg)
bootstrap_unnest_tbl(tb)
#> # A tibble: 50,000 × 2
#>    sim_number     y
#>    <fct>      <dbl>
#>  1 1           14.3
#>  2 1           15.5
#>  3 1           32.4
#>  4 1           13.3
#>  5 1           22.8
#>  6 1           21  
#>  7 1           24.4
#>  8 1           15.2
#>  9 1           26  
#> 10 1           27.3
#> # … with 49,990 more rows
#> # ℹ Use `print(n = ...)` to see more rows

bootstrap_unnest_tbl(tb) %>%
  tidy_distribution_summary_tbl(sim_number)
#> # A tibble: 2,000 × 13
#>    sim_num…¹ mean_…² media…³ std_val min_val max_val skewn…⁴ kurto…⁵ range   iqr
#>    <fct>       <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>
#>  1 1            21.0    19.7    6.53    10.4    33.9   0.543    2.38  23.5 10.5 
#>  2 2            19.6    19.7    4.33    13.3    30.4   0.497    2.85  17.1  7   
#>  3 3            20.7    19.2    5.45    10.4    32.4   0.908    3.40  22    3.60
#>  4 4            19.7    19.7    6.65    10.4    33.9   0.415    2.57  23.5  8.1 
#>  5 5            18.4    17.3    5.86    10.4    32.4   0.764    3.54  22    6.5 
#>  6 6            19.2    18.1    6.16    10.4    33.9   0.842    3.43  23.5  5.5 
#>  7 7            20.7    19.7    4.87    10.4    32.4   0.554    3.49  22    5   
#>  8 8            19.6    18.7    5.54    10.4    32.4   0.383    2.67  22    6.4 
#>  9 9            18.9    18.7    5.38    10.4    30.4   0.378    2.29  20    7.8 
#> 10 10           20.8    19.2    6.22    10.4    32.4   0.661    2.44  22    8.7 
#> # … with 1,990 more rows, 3 more variables: variance <dbl>, ci_low <dbl>,
#> #   ci_high <dbl>, and abbreviated variable names ¹​sim_number, ²​mean_val,
#> #   ³​median_val, ⁴​skewness, ⁵​kurtosis
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names