Tidy Randomly Generated Chisquare (Non-Central) Distribution Tibble

This function will generate n random points from a chisquare distribution with a user provided, .df, .ncp, and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

Usage

tidy_chisquare(.n = 50, .df = 1, .ncp = 1, .num_sims = 1)

Arguments

.n

The number of randomly generated points you want.

.df

Degrees of freedom (non-negative but can be non-integer)

.ncp

Non-centrality parameter, must be non-negative.

.num_sims

The number of randomly generated simulations you want.

Value

A tibble of randomly generated data.

Details

This function uses the underlying stats::rchisq(), and its underlying p, d, and q functions. For more information please see stats::rchisq()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_chisquare()
#> # A tibble: 50 × 7
#>    sim_number     x     y      dx       dy      p       q
#>    <fct>      <int> <dbl>   <dbl>    <dbl>  <dbl>   <dbl>
#>  1 1              1 0.341 -1.75   0.000989 0      0.00687
#>  2 1              2 1.25  -1.50   0.00352  0.0691 0.0926 
#>  3 1              3 1.16  -1.26   0.0107   0.0978 0.0792 
#>  4 1              4 1.06  -1.01   0.0277   0.120  0.0662 
#>  5 1              5 0.406 -0.767  0.0613   0.138  0.00977
#>  6 1              6 0.296 -0.523  0.116    0.155  0.00519
#>  7 1              7 0.348 -0.279  0.190    0.169  0.00719
#>  8 1              8 1.42  -0.0341 0.269    0.183  0.119  
#>  9 1              9 1.68   0.210  0.334    0.195  0.168  
#> 10 1             10 4.00   0.455  0.367    0.207  0.972  
#> # … with 40 more rows