Combine Multiple Tidy Distributions of Different Types

This allows a user to specify any n number of tidy_ distributions that can be combined into a single tibble. This is the preferred method for combining multiple distributions of different types, for example a Gaussian distribution and a Beta distribution.

This generates a single tibble with an added column of dist_type that will give the distribution family name and its associated parameters.

Usage

tidy_combine_distributions(...)

Arguments

...

The ... is where you can place your different distributions

Value

A tibble

Details

Allows a user to generate a tibble of different tidy_ distributions

See also

Author

Steven P. Sanderson II, MPH

Examples

tn <- tidy_normal()
tb <- tidy_beta()
tc <- tidy_cauchy()

tidy_combine_distributions(tn, tb, tc)
#> # A tibble: 150 × 8
#>    sim_number     x      y    dx       dy     p        q dist_type       
#>    <fct>      <int>  <dbl> <dbl>    <dbl> <dbl>    <dbl> <fct>           
#>  1 1              1 -0.503 -3.90 0.000237 0.5     -0.284 Gaussian c(0, 1)
#>  2 1              2 -2.22  -3.74 0.000788 0.508   -1.34  Gaussian c(0, 1)
#>  3 1              3 -0.254 -3.57 0.00221  0.516   -0.173 Gaussian c(0, 1)
#>  4 1              4  0.526 -3.41 0.00526  0.524    0.168 Gaussian c(0, 1)
#>  5 1              5 -0.266 -3.24 0.0108   0.533   -0.178 Gaussian c(0, 1)
#>  6 1              6 -0.921 -3.08 0.0193   0.541   -0.480 Gaussian c(0, 1)
#>  7 1              7 -0.184 -2.91 0.0310   0.549   -0.142 Gaussian c(0, 1)
#>  8 1              8 -2.74  -2.75 0.0452   0.557 -Inf     Gaussian c(0, 1)
#>  9 1              9  0.810 -2.58 0.0604   0.565    0.295 Gaussian c(0, 1)
#> 10 1             10  0.483 -2.42 0.0740   0.573    0.149 Gaussian c(0, 1)
#> # … with 140 more rows

## OR

tidy_combine_distributions(
  tidy_normal(),
  tidy_beta(),
  tidy_cauchy(),
  tidy_logistic()
)
#> # A tibble: 200 × 8
#>    sim_number     x       y    dx       dy     p        q dist_type       
#>    <fct>      <int>   <dbl> <dbl>    <dbl> <dbl>    <dbl> <fct>           
#>  1 1              1  1.68   -3.16 0.000279 0.5   Inf      Gaussian c(0, 1)
#>  2 1              2 -0.443  -3.04 0.000782 0.508  -0.127  Gaussian c(0, 1)
#>  3 1              3 -1.05   -2.92 0.00193  0.516  -0.549  Gaussian c(0, 1)
#>  4 1              4 -0.372  -2.80 0.00419  0.524  -0.0802 Gaussian c(0, 1)
#>  5 1              5 -0.415  -2.69 0.00807  0.533  -0.108  Gaussian c(0, 1)
#>  6 1              6  1.56   -2.57 0.0138   0.541   1.85   Gaussian c(0, 1)
#>  7 1              7  0.544  -2.45 0.0212   0.549   0.539  Gaussian c(0, 1)
#>  8 1              8 -0.180  -2.33 0.0292   0.557   0.0440 Gaussian c(0, 1)
#>  9 1              9 -0.791  -2.21 0.0367   0.565  -0.360  Gaussian c(0, 1)
#> 10 1             10 -0.0646 -2.09 0.0429   0.573   0.119  Gaussian c(0, 1)
#> # … with 190 more rows