Tidy Distribution Summary Statistics Tibble

This function returns a summary statistics tibble. It will use the y column from the tidy_ distribution function.

tidy_distribution_summary_tbl(.data, ...)

Arguments

.data

The data that is going to be passed from a a tidy_ distribution function.

...

This is the grouping variable that gets passed to dplyr::group_by() and dplyr::select().

Value

A summary stats tibble

Details

This function takes in a tidy_ distribution table and will return a tibble of the following information:

sim_number

mean_val

median_val

std_val

min_val

max_val

skewness

kurtosis

range

iqr

variance

The kurtosis and skewness come from the package healthyR.ai

Author

Steven P. Sanderson II, MPH

Examples

library(dplyr)
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

tn <- tidy_normal(.num_sims = 5)
tb <- tidy_beta(.num_sims = 5)

tidy_distribution_summary_tbl(tn)
#> # A tibble: 1 × 12
#>   mean_val median_…¹ std_val min_val max_val skewn…² kurto…³ range   iqr varia…⁴
#>      <dbl>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>   <dbl>
#> 1  -0.0168   -0.0468    1.05   -2.87    3.20 -0.0211    3.33  6.07  1.30    1.10
#> # … with 2 more variables: ci_low <dbl>, ci_high <dbl>, and abbreviated
#> #   variable names ¹​median_val, ²​skewness, ³​kurtosis, ⁴​variance
#> # ℹ Use `colnames()` to see all variable names
tidy_distribution_summary_tbl(tn, sim_number)
#> # A tibble: 5 × 13
#>   sim_number mean_…¹ media…² std_val min_val max_val skewn…³ kurto…⁴ range   iqr
#>   <fct>        <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>
#> 1 1          -0.192  -0.137    1.13    -2.87    2.56  -0.182    3.05  5.44 1.37 
#> 2 2           0.126   0.140    0.869   -2.45    2.04  -0.245    3.62  4.49 0.873
#> 3 3          -0.328  -0.391    0.982   -2.72    1.92   0.362    3.30  4.63 0.909
#> 4 4          -0.0644  0.0126   1.04    -2.50    2.41  -0.318    2.89  4.91 1.27 
#> 5 5           0.374   0.360    1.10    -2.53    3.20   0.190    3.50  5.73 1.42 
#> # … with 3 more variables: variance <dbl>, ci_low <dbl>, ci_high <dbl>, and
#> #   abbreviated variable names ¹​mean_val, ²​median_val, ³​skewness, ⁴​kurtosis
#> # ℹ Use `colnames()` to see all variable names

data_tbl <- tidy_combine_distributions(tn, tb)

tidy_distribution_summary_tbl(data_tbl)
#> # A tibble: 1 × 12
#>   mean_val median_…¹ std_val min_val max_val skewn…² kurto…³ range   iqr varia…⁴
#>      <dbl>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>   <dbl>
#> 1    0.254     0.384   0.816   -2.87    3.20  -0.786    5.17  6.07 0.782   0.666
#> # … with 2 more variables: ci_low <dbl>, ci_high <dbl>, and abbreviated
#> #   variable names ¹​median_val, ²​skewness, ³​kurtosis, ⁴​variance
#> # ℹ Use `colnames()` to see all variable names
tidy_distribution_summary_tbl(data_tbl, dist_type)
#> # A tibble: 2 × 13
#>   dist_type  mean_…¹ media…² std_val min_val max_val skewn…³ kurto…⁴ range   iqr
#>   <fct>        <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>
#> 1 Gaussian … -0.0168 -0.0468   1.05  -2.87      3.20 -0.0211    3.33 6.07  1.30 
#> 2 Beta c(1,…  0.525   0.586    0.292  0.0138    1.00 -0.195     1.83 0.986 0.473
#> # … with 3 more variables: variance <dbl>, ci_low <dbl>, ci_high <dbl>, and
#> #   abbreviated variable names ¹​mean_val, ²​median_val, ³​skewness, ⁴​kurtosis
#> # ℹ Use `colnames()` to see all variable names