Tidy Randomly Generated F Distribution Tibble

This function will generate n random points from a rf distribution with a user provided, df1,df2, and ncp, and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

tidy_f(.n = 50, .df1 = 1, .df2 = 1, .ncp = 0, .num_sims = 1)

Arguments

.n

The number of randomly generated points you want.

.df1

Degrees of freedom, Inf is allowed.

.df2

Degrees of freedom, Inf is allowed.

.ncp

Non-centrality parameter.

.num_sims

The number of randomly generated simulations you want.

Value

A tibble of randomly generated data.

Details

This function uses the underlying stats::rf(), and its underlying p, d, and q functions. For more information please see stats::rf()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_f()
#> # A tibble: 50 x 7
#>    sim_number     x       y     dx       dy     p       q
#>    <fct>      <int>   <dbl>  <dbl>    <dbl> <dbl>   <dbl>
#>  1 1              1 0.167   -8.46  0.000507     0 0      
#>  2 1              2 0.0200  -4.53  0.0151       0 0.00103
#>  3 1              3 0.0769  -0.594 0.0744       0 0.00412
#>  4 1              4 0.654    3.34  0.0695       0 0.00931
#>  5 1              5 1.52     7.27  0.0259       0 0.0166 
#>  6 1              6 0.580   11.2   0.0174       0 0.0261 
#>  7 1              7 0.0445  15.1   0.0103       0 0.0379 
#>  8 1              8 0.850   19.1   0.00479      0 0.0521 
#>  9 1              9 1.10    23.0   0.000762     0 0.0688 
#> 10 1             10 0.00582 26.9   0.000307     0 0.0881 
#> # ... with 40 more rows