Tidy Randomly Generated Gamma Distribution Tibble

This function will generate n random points from a gamma distribution with a user provided, .shape, .rate, and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

tidy_gamma(.n = 50, .shape = 1, .rate = 1, .num_sims = 1)

Arguments

.n

The number of randomly generated points you want.

.shape

This is strictly 0 to infinity.

.rate

The standard deviation of the randomly generated data. This is strictly from 0 to infinity.

.num_sims

The number of randomly generated simulations you want.

Value

A tibble of randomly generated data.

Details

This function uses the underlying stats::rgamma(), and its underlying p, d, and q functions. For more information please see stats::rgamma()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_gamma()
#> # A tibble: 50 x 7
#>    sim_number     x      y      dx      dy      p      q
#>    <fct>      <int>  <dbl>   <dbl>   <dbl>  <dbl>  <dbl>
#>  1 1              1 1.23   -0.870  0.00172 0      0     
#>  2 1              2 0.918  -0.721  0.00751 0.0202 0.0206
#>  3 1              3 0.0757 -0.571  0.0260  0.0400 0.0417
#>  4 1              4 0.133  -0.422  0.0720  0.0594 0.0632
#>  5 1              5 0.0835 -0.272  0.160   0.0784 0.0852
#>  6 1              6 0.466  -0.123  0.288   0.0970 0.108 
#>  7 1              7 1.22    0.0264 0.426   0.115  0.131 
#>  8 1              8 0.551   0.176  0.534   0.133  0.154 
#>  9 1              9 0.142   0.325  0.591   0.151  0.178 
#> 10 1             10 5.54    0.475  0.605   0.168  0.203 
#> # ... with 40 more rows