Tidy Randomly Generated Generalized Beta Distribution Tibble

This function will generate n random points from a generalized beta distribution with a user provided, .shape1, .shape2, .shape3, .rate, and/or .sclae, and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

tidy_generalized_beta(
  .n = 50,
  .shape1 = 1,
  .shape2 = 1,
  .shape3 = 1,
  .rate = 1,
  .scale = 1/.rate,
  .num_sims = 1
)

Arguments

.n

The number of randomly generated points you want.

.shape1

A non-negative parameter of the Beta distribution.

.shape2

A non-negative parameter of the Beta distribution.

.shape3

A non-negative parameter of the Beta distribution.

.rate

An alternative way to specify the .scale parameter.

.scale

Must be strictly positive.

.num_sims

The number of randomly generated simulations you want.

Value

A tibble of randomly generated data.

Details

This function uses the underlying stats::rbeta(), and its underlying p, d, and q functions. For more information please see stats::rbeta()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_generalized_beta()
#> # A tibble: 50 x 7
#>    sim_number     x       y      dx      dy      p      q
#>    <fct>      <int>   <dbl>   <dbl>   <dbl>  <dbl>  <dbl>
#>  1 1              1 0.175   -0.366  0.00167 0      0     
#>  2 1              2 0.00384 -0.331  0.00402 0.0204 0.0204
#>  3 1              3 0.255   -0.295  0.00899 0.0408 0.0408
#>  4 1              4 0.300   -0.260  0.0187  0.0612 0.0612
#>  5 1              5 0.768   -0.224  0.0361  0.0816 0.0816
#>  6 1              6 0.536   -0.189  0.0650  0.102  0.102 
#>  7 1              7 0.185   -0.154  0.109   0.122  0.122 
#>  8 1              8 0.460   -0.118  0.172   0.143  0.143 
#>  9 1              9 0.643   -0.0827 0.254   0.163  0.163 
#> 10 1             10 0.541   -0.0473 0.354   0.184  0.184 
#> # ... with 40 more rows