Tidy Randomly Generated Inverse Burr Distribution Tibble

This function will generate n random points from an Inverse Burr distribution with a user provided, .shape1, .shape2, .scale, .rate, and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

Usage

tidy_inverse_burr(
  .n = 50,
  .shape1 = 1,
  .shape2 = 1,
  .rate = 1,
  .scale = 1/.rate,
  .num_sims = 1,
  .return_tibble = TRUE
)

Arguments

.n

The number of randomly generated points you want.

.shape1

Must be strictly positive.

.shape2

Must be strictly positive.

.rate

An alternative way to specify the .scale.

.scale

Must be strictly positive.

.num_sims

The number of randomly generated simulations you want.

.return_tibble

A logical value indicating whether to return the result as a tibble. Default is TRUE.

Value

A tibble of randomly generated data.

Details

This function uses the underlying actuar::rinvburr(), and its underlying p, d, and q functions. For more information please see actuar::rinvburr()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_inverse_burr()
#> # A tibble: 50 × 7
#>    sim_number     x      y     dx      dy      p      q
#>    <fct>      <int>  <dbl>  <dbl>   <dbl>  <dbl>  <dbl>
#>  1 1              1 1.75   -2.93  0.00105 0.636  1.75  
#>  2 1              2 0.0532 -2.14  0.00954 0.0505 0.0532
#>  3 1              3 3.41   -1.34  0.0470  0.773  3.41  
#>  4 1              4 7.56   -0.550 0.128   0.883  7.56  
#>  5 1              5 0.103   0.243 0.203   0.0933 0.103 
#>  6 1              6 4.06    1.04  0.205   0.802  4.06  
#>  7 1              7 4.91    1.83  0.160   0.831  4.91  
#>  8 1              8 0.0668  2.62  0.119   0.0626 0.0668
#>  9 1              9 8.89    3.42  0.0919  0.899  8.89  
#> 10 1             10 4.90    4.21  0.0719  0.831  4.90  
#> # ℹ 40 more rows