Tidy Randomly Generated Lognormal Distribution Tibble

This function will generate n random points from a lognormal distribution with a user provided, .meanlog, .sdlog, and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

tidy_lognormal(.n = 50, .meanlog = 0, .sdlog = 1, .num_sims = 1)

Arguments

.n

The number of randomly generated points you want.

.meanlog

Mean of the distribution on the log scale with default 0

.sdlog

Standard deviation of the distribution on the log scale with default 1

.num_sims

The number of randomly generated simulations you want.

Value

A tibble of randomly generated data.

Details

This function uses the underlying stats::rlnorm(), and its underlying p, d, and q functions. For more information please see stats::rlnorm()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_lognormal()
#> # A tibble: 50 x 7
#>    sim_number     x     y      dx      dy         p     q
#>    <fct>      <int> <dbl>   <dbl>   <dbl>     <dbl> <dbl>
#>  1 1              1 0.232 -1.07   0.00141 0         0    
#>  2 1              2 0.836 -0.878  0.00555 0.0000497 0.129
#>  3 1              3 2.16  -0.687  0.0179  0.000690  0.175
#>  4 1              4 0.998 -0.497  0.0476  0.00261   0.213
#>  5 1              5 1.71  -0.307  0.105   0.00611   0.248
#>  6 1              6 0.170 -0.117  0.194   0.0112    0.281
#>  7 1              7 1.06   0.0732 0.303   0.0179    0.313
#>  8 1              8 0.596  0.263  0.407   0.0258    0.344
#>  9 1              9 0.870  0.453  0.476   0.0350    0.375
#> 10 1             10 1.37   0.644  0.499   0.0451    0.406
#> # ... with 40 more rows