Tidy Randomly Generated Gaussian Tibble

This function will generate n randong points from a gaussian distribution with a user provided, mean, standard deviation and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the dnorm, pnorm and qnorm data points as well.

The data is returned un-grouped.

tidy_normal(.n = 50, .mean = 0, .sd = 1, .num_sims = 1)

Arguments

.n

The number of randomly generated points you want.

.mean

The mean of the randomly generated data.

.sd

The standard deviation of the randomly generated data.

.num_sims

The number of randomly generated simulations you want.

Value

A tibble of randomly generated.

Details

This function uses the underlying stats::rnorm(), stats::pnorm(), and stats::qnorm() functions to generate data from the given parameters.

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_normal()
#> # A tibble: 50 x 6
#>    sim_number     x      y d_norm    p_norm   q_norm
#>    <fct>      <int>  <dbl>  <dbl>     <dbl>    <dbl>
#>  1 1              1  0.351 0.375  0         -Inf    
#>  2 1              2 -1.41  0.148  0           -2.05 
#>  3 1              3  1.75  0.0870 0           -1.74 
#>  4 1              4  1.47  0.135  0           -1.54 
#>  5 1              5  0.942 0.256  0           -1.39 
#>  6 1              6 -0.191 0.392  0           -1.27 
#>  7 1              7 -0.418 0.366  0           -1.16 
#>  8 1              8  0.286 0.383  4.18e-284   -1.07 
#>  9 1              9 -0.297 0.382  1.11e-253   -0.981
#> 10 1             10  0.441 0.362  5.45e-225   -0.901
#> # ... with 40 more rows