Tidy Randomly Generated Pareto Distribution Tibble

This function will generate n random points from a pareto distribution with a user provided, .shape, .scale, and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

tidy_pareto(.n = 50, .shape = 10, .scale = 0.1, .num_sims = 1)

Arguments

.n

The number of randomly generated points you want.

.shape

Must be positive.

.scale

Must be positive.

.num_sims

The number of randomly generated simulations you want.

Value

A tibble of randomly generated data.

Details

This function uses the underlying actuar::rpareto(), and its underlying p, d, and q functions. For more information please see actuar::rpareto()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_pareto()
#> # A tibble: 50 x 7
#>    sim_number     x        y        dx     dy     p        q
#>    <fct>      <int>    <dbl>     <dbl>  <dbl> <dbl>    <dbl>
#>  1 1              1 0.0120   -0.0112    0.136     0 0       
#>  2 1              2 0.000982 -0.00994   0.394     0 0.000206
#>  3 1              3 0.0347   -0.00864   1.02      0 0.000418
#>  4 1              4 0.00548  -0.00734   2.37      0 0.000634
#>  5 1              5 0.00672  -0.00604   4.94      0 0.000855
#>  6 1              6 0.0306   -0.00474   9.30      0 0.00108 
#>  7 1              7 0.00191  -0.00343  15.8       0 0.00131 
#>  8 1              8 0.00272  -0.00213  24.3       0 0.00155 
#>  9 1              9 0.00142  -0.000834 34.0       0 0.00180 
#> 10 1             10 0.00279   0.000467 43.5       0 0.00205 
#> # ... with 40 more rows