Tidy Randomly Generated Pareto Single Parameter Distribution Tibble

This function will generate n random points from a single parameter pareto distribution with a user provided, .shape, .min, and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

Usage

tidy_pareto1(
  .n = 50,
  .shape = 1,
  .min = 1,
  .num_sims = 1,
  .return_tibble = TRUE
)

Arguments

.n

The number of randomly generated points you want.

.shape

Must be positive.

.min

The lower bound of the support of the distribution.

.num_sims

The number of randomly generated simulations you want.

.return_tibble

A logical value indicating whether to return the result as a tibble. Default is TRUE.

Value

A tibble of randomly generated data.

Details

This function uses the underlying actuar::rpareto1(), and its underlying p, d, and q functions. For more information please see actuar::rpareto1()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_pareto1()
#> # A tibble: 50 × 7
#>    sim_number     x     y        dx      dy      p     q
#>    <fct>      <int> <dbl>     <dbl>   <dbl>  <dbl> <dbl>
#>  1 1              1  4.77 -1.17     0.00146 0.790   4.77
#>  2 1              2  3.63 -0.586    0.0131  0.725   3.63
#>  3 1              3  1.20  0.000728 0.0650  0.165   1.20
#>  4 1              4  3.56  0.588    0.181   0.719   3.56
#>  5 1              5  1.18  1.17     0.289   0.152   1.18
#>  6 1              6  1.62  1.76     0.282   0.381   1.62
#>  7 1              7  4.01  2.35     0.194   0.750   4.01
#>  8 1              8  1.61  2.94     0.136   0.380   1.61
#>  9 1              9 18.2   3.52     0.128   0.945  18.2 
#> 10 1             10  1.09  4.11     0.116   0.0847  1.09
#> # ℹ 40 more rows