Compute Skewness of a Vector

This function takes in a vector as it's input and will return the skewness of that vector. The length of this vector must be at least four numbers. The skewness explains the 'tailedness' of the distribution of data.

((1/n) * sum(x - mu})^3) / ((()1/n) * sum(x - mu)^2)^(3/2)

tidy_skewness_vec(.x)

Arguments

.x

A numeric vector of length four or more.

Value

The skewness of a vector

Details

A function to return the skewness of a vector.

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_skewness_vec(rnorm(100, 3, 2))
#> [1] 0.05709072