Tidy Randomly Generated T Distribution Tibble

This function will generate n random points from a rt distribution with a user provided, df, ncp, and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

tidy_t(.n = 50, .df = 1, .ncp = 0, .num_sims = 1)

Arguments

.n

The number of randomly generated points you want.

.df

Degrees of freedom, Inf is allowed.

.ncp

Non-centrality parameter.

.num_sims

The number of randomly generated simulations you want.

Value

A tibble of randomly generated data.

Details

This function uses the underlying stats::rt(), and its underlying p, d, and q functions. For more information please see stats::rt()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_t()
#> # A tibble: 50 × 7
#>    sim_number     x        y     dx       dy     p       q
#>    <fct>      <int>    <dbl>  <dbl>    <dbl> <dbl>   <dbl>
#>  1 1              1  -0.211  -31.9  1.70e- 4 0.5     -1.27
#>  2 1              2   4.30   -29.0  5.68e- 3 0.506   -1.03
#>  3 1              3 -29.8    -26.1  1.60e- 5 0.513 -Inf   
#>  4 1              4   0.227  -23.1  1.00e- 2 0.519   -1.24
#>  5 1              5 110.     -20.2  5.07e- 7 0.526  Inf   
#>  6 1              6   2.12   -17.3  0        0.532   -1.14
#>  7 1              7  -0.0181 -14.3  8.72e-18 0.539   -1.26
#>  8 1              8  -0.112  -11.4  0        0.545   -1.26
#>  9 1              9   2.48    -8.47 2.90e-15 0.552   -1.12
#> 10 1             10  -2.63    -5.54 2.76e- 5 0.558   -1.42
#> # … with 40 more rows
#> # ℹ Use `print(n = ...)` to see more rows