Tidy Randomly Generated Uniform Distribution Tibble

This function will generate n random points from a uniform distribution with a user provided, .min and .max values, and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

Usage

tidy_uniform(.n = 50, .min = 0, .max = 1, .num_sims = 1)

Arguments

.n

The number of randomly generated points you want.

.min

A lower limit of the distribution.

.max

An upper limit of the distribution

.num_sims

The number of randomly generated simulations you want.

Value

A tibble of randomly generated data.

Details

This function uses the underlying stats::runif(), and its underlying p, d, and q functions. For more information please see stats::runif()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_uniform()
#> # A tibble: 50 × 7
#>    sim_number     x      y      dx      dy      p      q
#>    <fct>      <int>  <dbl>   <dbl>   <dbl>  <dbl>  <dbl>
#>  1 1              1 0.321  -0.345  0.00337 0.321  0.321 
#>  2 1              2 0.977  -0.311  0.00803 0.977  0.977 
#>  3 1              3 0.0515 -0.277  0.0177  0.0515 0.0515
#>  4 1              4 0.0632 -0.243  0.0359  0.0632 0.0632
#>  5 1              5 0.796  -0.209  0.0676  0.796  0.796 
#>  6 1              6 0.910  -0.174  0.118   0.910  0.910 
#>  7 1              7 0.607  -0.140  0.192   0.607  0.607 
#>  8 1              8 0.668  -0.106  0.290   0.668  0.668 
#>  9 1              9 0.767  -0.0717 0.411   0.767  0.767 
#> 10 1             10 0.681  -0.0375 0.546   0.681  0.681 
#> # … with 40 more rows