Tidy Randomly Generated Binomial Distribution Tibble

This function will generate n random points from a zero truncated binomial distribution with a user provided, .size, .prob, and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

tidy_zero_truncated_negative_binomial(
  .n = 50,
  .size = 0,
  .prob = 1,
  .num_sims = 1
)

Arguments

.n

The number of randomly generated points you want.

.size

Number of trials, zero or more.

.prob

Probability of success on each trial 0 <= prob <= 1.

.num_sims

The number of randomly generated simulations you want.

Value

A tibble of randomly generated data.

Details

This function uses the underlying actuar::rztnbinom(), and its underlying p, d, and q functions. For more information please see actuar::rztnbinom()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_zero_truncated_binomial()
#> Warning: NaNs produced
#> Warning: NaNs produced
#> # A tibble: 50 x 7
#>    sim_number     x     y     dx     dy     p     q
#>    <fct>      <int> <int>  <dbl>  <dbl> <dbl> <dbl>
#>  1 1              1     0 -1.23  0.0109   NaN   NaN
#>  2 1              2     0 -1.18  0.0156   NaN   NaN
#>  3 1              3     0 -1.13  0.0220   NaN   NaN
#>  4 1              4     0 -1.08  0.0305   NaN   NaN
#>  5 1              5     0 -1.03  0.0418   NaN   NaN
#>  6 1              6     0 -0.983 0.0564   NaN   NaN
#>  7 1              7     0 -0.932 0.0749   NaN   NaN
#>  8 1              8     0 -0.882 0.0981   NaN   NaN
#>  9 1              9     0 -0.832 0.126    NaN   NaN
#> 10 1             10     0 -0.781 0.161    NaN   NaN
#> # ... with 40 more rows