Calculate Akaike Information Criterion (AIC) for Cauchy Distribution

This function estimates the parameters of a Cauchy distribution from the provided data using maximum likelihood estimation, and then calculates the AIC value based on the fitted distribution.

Usage

util_cauchy_aic(.x)

Arguments

.x

A numeric vector containing the data to be fitted to a Cauchy distribution.

Value

The AIC value calculated based on the fitted Cauchy distribution to the provided data.

Details

This function calculates the Akaike Information Criterion (AIC) for a Cauchy distribution fitted to the provided data.

This function fits a Cauchy distribution to the provided data using maximum likelihood estimation. It first estimates the initial parameters of the Cauchy distribution using the method of moments. Then, it optimizes the negative log-likelihood function using the provided data and the initial parameter estimates. Finally, it calculates the AIC value based on the fitted distribution.

Initial parameter estimates: The function uses the method of moments estimates for the initial location and scale parameters of the Cauchy distribution.

Optimization method: The function uses the optim function for optimization. You might explore different optimization methods within optim for potentially better performance.

Goodness-of-fit: While AIC is a useful metric for model comparison, it's recommended to also assess the goodness-of-fit of the chosen model using visualization and other statistical tests.

See also

Author

Steven P. Sanderson II, MPH

Examples

# Example 1: Calculate AIC for a sample dataset
set.seed(123)
x <- rcauchy(30)
util_cauchy_aic(x)
#> [1] 152.304