Distribution Statistics

Returns distribution statistics in a tibble.

util_chisquare_stats_tbl(.data)

Arguments

.data

The data being passed from a tidy_ distribution function.

Value

A tibble

Details

This function will take in a tibble and returns the statistics of the given type of tidy_ distribution. It is required that data be passed from a tidy_ distribution function.

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_chisquare() %>%
  util_chisquare_stats_tbl()
#> # A tibble: 1 x 15
#>   tidy_function  function_call  distribution distribution_ty~ points simulations
#>   <chr>          <chr>          <chr>        <chr>             <dbl>       <dbl>
#> 1 tidy_chisquare Chisquare c(1~ Chisquare    continuous           50           1
#> # ... with 9 more variables: mean <dbl>, median <dbl>, mode <chr>,
#> #   std_dv <dbl>, coeff_var <dbl>, skewness <dbl>, kurtosis <dbl>,
#> #   computed_std_skew <dbl>, computed_std_kurt <dbl>