Distribution Statistics

Returns distribution statistics in a tibble.

Usage

util_exponential_stats_tbl(.data)

Arguments

.data

The data being passed from a tidy_ distribution function.

Value

A tibble

Details

This function will take in a tibble and returns the statistics of the given type of tidy_ distribution. It is required that data be passed from a tidy_ distribution function.

See also

Author

Steven P. Sanderson II, MPH

Examples

library(dplyr)

tidy_exponential() |>
  util_exponential_stats_tbl() |>
  glimpse()
#> Rows: 1
#> Columns: 18
#> $ tidy_function     <chr> "tidy_exponential"
#> $ function_call     <chr> "Exponential c(1)"
#> $ distribution      <chr> "Exponential"
#> $ distribution_type <chr> "continuous"
#> $ points            <dbl> 50
#> $ simulations       <dbl> 1
#> $ mean              <dbl> 1
#> $ median            <dbl> 0.6931472
#> $ mode              <dbl> 1
#> $ range             <chr> "1 to Inf"
#> $ std_dv            <dbl> 1
#> $ coeff_var         <dbl> 1
#> $ skewness          <dbl> 2
#> $ kurtosis          <dbl> 9
#> $ computed_std_skew <dbl> 1.42207
#> $ computed_std_kurt <dbl> 4.940398
#> $ ci_lo             <dbl> 0.009799229
#> $ ci_hi             <dbl> 3.545823