Calculate Akaike Information Criterion (AIC) for Pareto Distribution

This function estimates the shape and scale parameters of a Pareto distribution from the provided data using maximum likelihood estimation, and then calculates the AIC value based on the fitted distribution.

Usage

util_pareto1_aic(.x)

Arguments

.x

A numeric vector containing the data to be fitted to a Pareto distribution.

Value

The AIC value calculated based on the fitted Pareto distribution to the provided data.

Details

This function calculates the Akaike Information Criterion (AIC) for a Pareto distribution fitted to the provided data.

This function fits a Pareto distribution to the provided data using maximum likelihood estimation. It estimates the shape and scale parameters of the Pareto distribution using maximum likelihood estimation. Then, it calculates the AIC value based on the fitted distribution.

Initial parameter estimates: The function uses the method of moments estimates as starting points for the shape and scale parameters of the Pareto distribution.

Optimization method: The function uses the optim function for optimization. You might explore different optimization methods within optim for potentially better performance.

Goodness-of-fit: While AIC is a useful metric for model comparison, it's recommended to also assess the goodness-of-fit of the chosen model using visualization and other statistical tests.

See also

Author

Steven P. Sanderson II, MPH

Examples

# Example 1: Calculate AIC for a sample dataset
set.seed(123)
x <- tidy_pareto1()$y
util_pareto1_aic(x)
#> [1] 185.0364