Calculate Akaike Information Criterion (AIC) for Zero-Truncated Geometric Distribution

This function estimates the probability parameter of a Zero-Truncated Geometric distribution from the provided data and calculates the AIC value based on the fitted distribution.

Usage

util_zero_truncated_geometric_aic(.x)

Arguments

.x

A numeric vector containing the data to be fitted to a Zero-Truncated Geometric distribution.

Value

The AIC value calculated based on the fitted Zero-Truncated Geometric distribution to the provided data.

Details

This function calculates the Akaike Information Criterion (AIC) for a Zero-Truncated Geometric distribution fitted to the provided data.

This function fits a Zero-Truncated Geometric distribution to the provided data. It estimates the probability parameter using the method of moments and calculates the AIC value.

Optimization method: Since the parameter is directly calculated from the data, no optimization is needed.

Goodness-of-fit: While AIC is a useful metric for model comparison, it's recommended to also assess the goodness-of-fit of the chosen model using visualization and other statistical tests.

See also

Author

Steven P. Sanderson II, MPH

Examples

library(actuar)
#> 
#> Attaching package: 'actuar'
#> The following objects are masked from 'package:stats':
#> 
#>     sd, var
#> The following object is masked from 'package:grDevices':
#> 
#>     cm

# Example: Calculate AIC for a sample dataset
set.seed(123)
x <- rztgeom(100, prob = 0.2)
util_zero_truncated_geometric_aic(x)
#> [1] 492.3338