Augment Bootstrap Q

Takes a numeric vector and will return the quantile.

Usage

bootstrap_q_augment(.data, .value, .names = "auto")

Arguments

.data

The data being passed that will be augmented by the function.

.value

This is passed rlang::enquo() to capture the vectors you want to augment.

.names

The default is "auto"

Value

A augmented tibble

Details

Takes a numeric vector and will return the quantile of that vector. This function is intended to be used on its own in order to add columns to a tibble.

See also

Author

Steven P. Sanderson II, MPH

Examples

x <- mtcars$mpg

tidy_bootstrap(x) %>%
  bootstrap_unnest_tbl() %>%
  bootstrap_q_augment(y)
#> # A tibble: 50,000 × 3
#>    sim_number     y     q
#>    <fct>      <dbl> <dbl>
#>  1 1           15.8  10.4
#>  2 1           30.4  10.4
#>  3 1           21.5  10.4
#>  4 1           14.7  10.4
#>  5 1           15    10.4
#>  6 1           14.7  10.4
#>  7 1           16.4  10.4
#>  8 1           21.4  10.4
#>  9 1           30.4  10.4
#> 10 1           22.8  10.4
#> # … with 49,990 more rows