Unnest Tidy Bootstrap Tibble

Unnest the data output from tidy_bootstrap().

bootstrap_unnest_tbl(.data)

Arguments

.data

The data that is passed from the tidy_bootstrap() function.

Value

A tibble

Details

This function takes as input the output of the tidy_bootstrap() function and returns a two column tibble. The columns are sim_number and y

It looks for an attribute that comes from using tidy_bootstrap() so it will not work unless the data comes from that function.

See also

Author

Steven P. Sanderson II, MPH

Examples

tb <- tidy_bootstrap(.x = mtcars$mpg)
bootstrap_unnest_tbl(tb)
#> # A tibble: 50,000 x 2
#>    sim_number     y
#>    <fct>      <dbl>
#>  1 1           14.3
#>  2 1           19.2
#>  3 1           13.3
#>  4 1           21.4
#>  5 1           16.4
#>  6 1           17.3
#>  7 1           15.2
#>  8 1           30.4
#>  9 1           15  
#> 10 1           18.7
#> # ... with 49,990 more rows

bootstrap_unnest_tbl(tb) %>%
  tidy_distribution_summary_tbl(sim_number)
#> # A tibble: 2,000 x 11
#>    sim_number mean_val median_val std_val min_val max_val skewness kurtosis
#>    <fct>         <dbl>      <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>
#>  1 1              18.3       17.3    4.66    10.4    30.4    1.24      4.58
#>  2 2              21.8       21      6.45    10.4    33.9    0.353     2.21
#>  3 3              20.9       21      6.77    10.4    33.9    0.367     2.17
#>  4 4              18.0       17.3    6.48    10.4    32.4    0.868     3.15
#>  5 5              18.1       16.4    5.98    10.4    33.9    1.05      3.99
#>  6 6              20.1       19.2    6.12    10.4    33.9    0.852     2.97
#>  7 7              20.5       21      6.18    10.4    33.9    0.342     2.46
#>  8 8              19.2       17.8    6.01    10.4    33.9    1.16      3.56
#>  9 9              17.4       16.4    4.97    10.4    30.4    0.618     3.27
#> 10 10             21.8       21      6.29    14.3    32.4    0.364     1.60
#> # ... with 1,990 more rows, and 3 more variables: range <dbl>, iqr <dbl>,
#> #   variance <dbl>