Unnest Tidy Bootstrap Tibble

Unnest the data output from tidy_bootstrap().

Usage

bootstrap_unnest_tbl(.data)

Arguments

.data

The data that is passed from the tidy_bootstrap() function.

Value

A tibble

Details

This function takes as input the output of the tidy_bootstrap() function and returns a two column tibble. The columns are sim_number and y

It looks for an attribute that comes from using tidy_bootstrap() so it will not work unless the data comes from that function.

See also

Author

Steven P. Sanderson II, MPH

Examples

tb <- tidy_bootstrap(.x = mtcars$mpg)
bootstrap_unnest_tbl(tb)
#> # A tibble: 50,000 × 2
#>    sim_number     y
#>    <fct>      <dbl>
#>  1 1           30.4
#>  2 1           16.4
#>  3 1           10.4
#>  4 1           18.7
#>  5 1           17.8
#>  6 1           21.4
#>  7 1           10.4
#>  8 1           10.4
#>  9 1           32.4
#> 10 1           26  
#> # … with 49,990 more rows

bootstrap_unnest_tbl(tb) %>%
  tidy_distribution_summary_tbl(sim_number)
#> # A tibble: 2,000 × 13
#>    sim_num…¹ mean_…² media…³ std_val min_val max_val skewn…⁴ kurto…⁵ range   iqr
#>    <fct>       <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>
#>  1 1            20.6    19.2    6.62    10.4    32.4   0.277    2.23  22     9.6
#>  2 2            19.7    19.2    5.80    10.4    33.9   0.721    3.45  23.5   5.2
#>  3 3            19.2    18.7    5.89    10.4    33.9   1.07     3.87  23.5   7.6
#>  4 4            22.2    19.7    6.08    14.3    33.9   0.544    1.98  19.6   8.2
#>  5 5            20.9    18.7    7.25    10.4    33.9   0.618    2.30  23.5  10.5
#>  6 6            19.4    18.1    6.46    10.4    33.9   0.896    3.12  23.5   6.3
#>  7 7            20.6    18.7    6.42    10.4    33.9   0.679    2.55  23.5   8.9
#>  8 8            19.2    17.3    6.05    10.4    33.9   0.981    3.04  23.5   6.2
#>  9 9            22.3    21.4    6.20    10.4    33.9   0.385    2.40  23.5   5.7
#> 10 10           20.2    19.2    5.99    10.4    33.9   1.04     3.62  23.5   5.1
#> # … with 1,990 more rows, 3 more variables: variance <dbl>, ci_low <dbl>,
#> #   ci_high <dbl>, and abbreviated variable names ¹​sim_number, ²​mean_val,
#> #   ³​median_val, ⁴​skewness, ⁵​kurtosis