Unnest Tidy Bootstrap Tibble

Unnest the data output from tidy_bootstrap().

Usage

bootstrap_unnest_tbl(.data)

Arguments

.data

The data that is passed from the tidy_bootstrap() function.

Value

A tibble

Details

This function takes as input the output of the tidy_bootstrap() function and returns a two column tibble. The columns are sim_number and y

It looks for an attribute that comes from using tidy_bootstrap() so it will not work unless the data comes from that function.

See also

Author

Steven P. Sanderson II, MPH

Examples

tb <- tidy_bootstrap(.x = mtcars$mpg)
bootstrap_unnest_tbl(tb)
#> # A tibble: 50,000 × 2
#>    sim_number     y
#>    <fct>      <dbl>
#>  1 1           19.2
#>  2 1           15.2
#>  3 1           15.8
#>  4 1           14.7
#>  5 1           22.8
#>  6 1           30.4
#>  7 1           21.5
#>  8 1           14.7
#>  9 1           22.8
#> 10 1           21  
#> # … with 49,990 more rows
#> # ℹ Use `print(n = ...)` to see more rows

bootstrap_unnest_tbl(tb) %>%
  tidy_distribution_summary_tbl(sim_number)
#> # A tibble: 2,000 × 13
#>    sim_num…¹ mean_…² media…³ std_val min_val max_val skewn…⁴ kurto…⁵ range   iqr
#>    <fct>       <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>
#>  1 1            18.5    18.7    4.48    10.4    30.4  0.406     3.46  20     6.3
#>  2 2            18.8    17.8    4.95    10.4    30.4  0.796     3.67  20     5.9
#>  3 3            20.3    19.2    6.30    10.4    33.9  0.744     2.64  23.5   7.6
#>  4 4            21.1    21.4    6.46    10.4    32.4 -0.0897    2.15  22    10.5
#>  5 5            21.5    21      5.20    13.3    32.4  0.618     2.82  19.1   3.6
#>  6 6            19.3    19.2    5.75    10.4    32.4  0.357     2.74  22     7.8
#>  7 7            20.4    18.7    6.60    10.4    33.9  0.614     2.29  23.5   9.2
#>  8 8            16.8    17.8    4.60    10.4    30.4  0.629     4.30  20     4.9
#>  9 9            21.1    19.2    6.49    10.4    33.9  0.416     2.04  23.5  10.2
#> 10 10           22.4    21.4    4.48    14.7    33.9  0.490     3.46  19.2   3.4
#> # … with 1,990 more rows, 3 more variables: variance <dbl>, ci_low <dbl>,
#> #   ci_high <dbl>, and abbreviated variable names ¹​sim_number, ²​mean_val,
#> #   ³​median_val, ⁴​skewness, ⁵​kurtosis
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names