Unnest Tidy Bootstrap Tibble

Unnest the data output from tidy_bootstrap().

Usage

bootstrap_unnest_tbl(.data)

Arguments

.data

The data that is passed from the tidy_bootstrap() function.

Value

A tibble

Details

This function takes as input the output of the tidy_bootstrap() function and returns a two column tibble. The columns are sim_number and y

It looks for an attribute that comes from using tidy_bootstrap() so it will not work unless the data comes from that function.

See also

Author

Steven P. Sanderson II, MPH

Examples

tb <- tidy_bootstrap(.x = mtcars$mpg)
bootstrap_unnest_tbl(tb)
#> # A tibble: 50,000 × 2
#>    sim_number     y
#>    <fct>      <dbl>
#>  1 1           17.3
#>  2 1           15.2
#>  3 1           17.3
#>  4 1           27.3
#>  5 1           18.1
#>  6 1           14.7
#>  7 1           30.4
#>  8 1           21  
#>  9 1           30.4
#> 10 1           32.4
#> # ℹ 49,990 more rows

bootstrap_unnest_tbl(tb) %>%
  tidy_distribution_summary_tbl(sim_number)
#> # A tibble: 2,000 × 13
#>    sim_number mean_val median_val std_val min_val max_val skewness kurtosis
#>    <fct>         <dbl>      <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>
#>  1 1              19.7       18.1    5.95    10.4    32.4   0.581      2.65
#>  2 2              18.7       19.7    4.58    10.4    30.4   0.139      3.24
#>  3 3              19.6       18.7    5.19    10.4    32.4   0.785      3.27
#>  4 4              19.9       17.3    6.07    10.4    30.4   0.654      2.18
#>  5 5              19.0       19.2    4.92    10.4    30.4   0.517      3.50
#>  6 6              19.1       18.7    4.90    10.4    30.4   0.204      2.75
#>  7 7              19.2       19.2    6.74    10.4    32.4   0.404      2.37
#>  8 8              20.9       19.7    5.91    10.4    33.9   0.656      2.91
#>  9 9              19.6       19.7    4.04    10.4    30.4   0.0864     4.03
#> 10 10             19.0       17.8    6.23    10.4    32.4   0.599      2.63
#> # ℹ 1,990 more rows
#> # ℹ 5 more variables: range <dbl>, iqr <dbl>, variance <dbl>, ci_low <dbl>,
#> #   ci_high <dbl>