Combine Multiple Tidy Distributions of Different Types

This allows a user to specify any n number of tidy_ distributions that can be combined into a single tibble. This is the preferred method for combining multiple distributions of different types, for example a Gaussian distribution and a Beta distribution.

This generates a single tibble with an added column of dist_type that will give the distribution family name and its associated parameters.

Usage

tidy_combine_distributions(...)

Arguments

...

The ... is where you can place your different distributions

Value

A tibble

Details

Allows a user to generate a tibble of different tidy_ distributions

See also

Author

Steven P. Sanderson II, MPH

Examples

tn <- tidy_normal()
tb <- tidy_beta()
tc <- tidy_cauchy()

tidy_combine_distributions(tn, tb, tc)
#> # A tibble: 150 × 8
#>    sim_number     x       y    dx       dy     p         q dist_type       
#>    <fct>      <int>   <dbl> <dbl>    <dbl> <dbl>     <dbl> <fct>           
#>  1 1              1 -0.0534 -3.42 0.000292 0.5     -0.0372 Gaussian c(0, 1)
#>  2 1              2 -0.182  -3.28 0.000991 0.508   -0.103  Gaussian c(0, 1)
#>  3 1              3 -0.557  -3.14 0.00282  0.516   -0.298  Gaussian c(0, 1)
#>  4 1              4 -2.44   -3.00 0.00677  0.524 -Inf      Gaussian c(0, 1)
#>  5 1              5 -1.09   -2.86 0.0138   0.533   -0.602  Gaussian c(0, 1)
#>  6 1              6 -1.49   -2.72 0.0241   0.541   -0.866  Gaussian c(0, 1)
#>  7 1              7 -1.68   -2.58 0.0369   0.549   -1.02   Gaussian c(0, 1)
#>  8 1              8 -1.02   -2.43 0.0509   0.557   -0.560  Gaussian c(0, 1)
#>  9 1              9  1.05   -2.29 0.0665   0.565    0.551  Gaussian c(0, 1)
#> 10 1             10 -1.52   -2.15 0.0864   0.573   -0.892  Gaussian c(0, 1)
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows

## OR

tidy_combine_distributions(
  tidy_normal(),
  tidy_beta(),
  tidy_cauchy(),
  tidy_logistic()
)
#> # A tibble: 200 × 8
#>    sim_number     x         y    dx       dy     p       q dist_type       
#>    <fct>      <int>     <dbl> <dbl>    <dbl> <dbl>   <dbl> <fct>           
#>  1 1              1  0.600    -4.54 0.000200 0.5    0.201  Gaussian c(0, 1)
#>  2 1              2 -2.21     -4.35 0.000625 0.508 -1.07   Gaussian c(0, 1)
#>  3 1              3 -1.31     -4.16 0.00167  0.516 -0.575  Gaussian c(0, 1)
#>  4 1              4 -0.000816 -3.97 0.00387  0.524 -0.0345 Gaussian c(0, 1)
#>  5 1              5 -0.289    -3.78 0.00776  0.533 -0.148  Gaussian c(0, 1)
#>  6 1              6 -2.00     -3.60 0.0136   0.541 -0.934  Gaussian c(0, 1)
#>  7 1              7 -0.104    -3.41 0.0210   0.549 -0.0749 Gaussian c(0, 1)
#>  8 1              8  0.827    -3.22 0.0293   0.557  0.293  Gaussian c(0, 1)
#>  9 1              9 -0.364    -3.03 0.0378   0.565 -0.177  Gaussian c(0, 1)
#> 10 1             10  0.897    -2.84 0.0468   0.573  0.321  Gaussian c(0, 1)
#> # … with 190 more rows
#> # ℹ Use `print(n = ...)` to see more rows