Combine Multiple Tidy Distributions of Different Types

This allows a user to specify any n number of tidy_ distributions that can be combined into a single tibble. This is the preferred method for combining multiple distributions of different types, for example a Gaussian distribution and a Beta distribution.

This generates a single tibble with an added column of dist_type that will give the distribution family name and its associated parameters.

Usage

tidy_combine_distributions(...)

Arguments

...

The ... is where you can place your different distributions

Value

A tibble

Details

Allows a user to generate a tibble of different tidy_ distributions

See also

Author

Steven P. Sanderson II, MPH

Examples

tn <- tidy_normal()
tb <- tidy_beta()
tc <- tidy_cauchy()

tidy_combine_distributions(tn, tb, tc)
#> # A tibble: 150 × 8
#>    sim_number     x      y    dx       dy     p      q dist_type       
#>    <fct>      <int>  <dbl> <dbl>    <dbl> <dbl>  <dbl> <fct>           
#>  1 1              1  0.773 -3.52 0.000233 0.5    0.482 Gaussian c(0, 1)
#>  2 1              2 -1.17  -3.38 0.000651 0.508 -0.640 Gaussian c(0, 1)
#>  3 1              3 -1.11  -3.24 0.00159  0.516 -0.600 Gaussian c(0, 1)
#>  4 1              4 -0.911 -3.10 0.00343  0.524 -0.475 Gaussian c(0, 1)
#>  5 1              5  0.226 -2.96 0.00650  0.533  0.165 Gaussian c(0, 1)
#>  6 1              6 -0.318 -2.82 0.0109   0.541 -0.134 Gaussian c(0, 1)
#>  7 1              7 -0.843 -2.68 0.0164   0.549 -0.433 Gaussian c(0, 1)
#>  8 1              8  0.934 -2.54 0.0227   0.557  0.584 Gaussian c(0, 1)
#>  9 1              9 -1.68  -2.40 0.0298   0.565 -1.04  Gaussian c(0, 1)
#> 10 1             10  1.54  -2.26 0.0389   0.573  1.05  Gaussian c(0, 1)
#> # … with 140 more rows
#> # ℹ Use `print(n = ...)` to see more rows

## OR

tidy_combine_distributions(
  tidy_normal(),
  tidy_beta(),
  tidy_cauchy(),
  tidy_logistic()
)
#> # A tibble: 200 × 8
#>    sim_number     x       y    dx       dy     p      q dist_type       
#>    <fct>      <int>   <dbl> <dbl>    <dbl> <dbl>  <dbl> <fct>           
#>  1 1              1  0.302  -4.27 0.000189 0.5    0.315 Gaussian c(0, 1)
#>  2 1              2  1.02   -4.11 0.000492 0.508  0.724 Gaussian c(0, 1)
#>  3 1              3 -0.745  -3.94 0.00115  0.516 -0.214 Gaussian c(0, 1)
#>  4 1              4 -1.26   -3.78 0.00240  0.524 -0.486 Gaussian c(0, 1)
#>  5 1              5  0.709  -3.62 0.00451  0.533  0.538 Gaussian c(0, 1)
#>  6 1              6 -1.23   -3.46 0.00767  0.541 -0.470 Gaussian c(0, 1)
#>  7 1              7  0.0644 -3.30 0.0119   0.549  0.193 Gaussian c(0, 1)
#>  8 1              8 -0.529  -3.14 0.0169   0.557 -0.105 Gaussian c(0, 1)
#>  9 1              9  0.413  -2.98 0.0224   0.565  0.374 Gaussian c(0, 1)
#> 10 1             10  1.73   -2.82 0.0283   0.573  1.32  Gaussian c(0, 1)
#> # … with 190 more rows
#> # ℹ Use `print(n = ...)` to see more rows