Combine Multiple Tidy Distributions of Different Types

This allows a user to specify any n number of tidy_ distributions that can be combined into a single tibble. This is the preferred method for combining multiple distributions of different types, for example a Gaussian distribution and a Beta distribution.

This generates a single tibble with an added column of dist_type that will give the distribution family name and its associated parameters.

tidy_combine_distributions(...)

Arguments

...

The ... is where you can place your different distributions

Value

A tibble

Details

Allows a user to generate a tibble of different tidy_ distributions

See also

Author

Steven P. Sanderson II, MPH

Examples

tn <- tidy_normal()
tb <- tidy_beta()
tc <- tidy_cauchy()

tidy_combine_distributions(tn, tb, tc)
#> # A tibble: 150 × 8
#>    sim_number     x       y    dx       dy     p      q dist_type       
#>    <fct>      <int>   <dbl> <dbl>    <dbl> <dbl>  <dbl> <fct>           
#>  1 1              1  0.256  -3.01 0.000342 0.5    0.237 Gaussian c(0, 1)
#>  2 1              2 -0.413  -2.89 0.000920 0.508 -0.221 Gaussian c(0, 1)
#>  3 1              3 -0.603  -2.78 0.00224  0.516 -0.356 Gaussian c(0, 1)
#>  4 1              4  0.634  -2.66 0.00495  0.524  0.513 Gaussian c(0, 1)
#>  5 1              5  0.376  -2.54 0.00994  0.533  0.321 Gaussian c(0, 1)
#>  6 1              6  0.0853 -2.42 0.0182   0.541  0.119 Gaussian c(0, 1)
#>  7 1              7  0.649  -2.30 0.0305   0.549  0.525 Gaussian c(0, 1)
#>  8 1              8  0.585  -2.18 0.0469   0.557  0.475 Gaussian c(0, 1)
#>  9 1              9  0.624  -2.06 0.0666   0.565  0.505 Gaussian c(0, 1)
#> 10 1             10  0.745  -1.94 0.0875   0.573  0.600 Gaussian c(0, 1)
#> # … with 140 more rows

## OR

tidy_combine_distributions(
  tidy_normal(),
  tidy_beta(),
  tidy_cauchy(),
  tidy_logistic()
)
#> # A tibble: 200 × 8
#>    sim_number     x       y    dx       dy     p       q dist_type       
#>    <fct>      <int>   <dbl> <dbl>    <dbl> <dbl>   <dbl> <fct>           
#>  1 1              1  0.412  -3.80 0.000212 0.5    0.147  Gaussian c(0, 1)
#>  2 1              2 -0.270  -3.64 0.000593 0.508 -0.186  Gaussian c(0, 1)
#>  3 1              3 -0.353  -3.48 0.00147  0.516 -0.227  Gaussian c(0, 1)
#>  4 1              4  1.60   -3.32 0.00322  0.524  0.801  Gaussian c(0, 1)
#>  5 1              5  0.150  -3.16 0.00625  0.533  0.0191 Gaussian c(0, 1)
#>  6 1              6  2.28   -3.00 0.0108   0.541  1.40   Gaussian c(0, 1)
#>  7 1              7  0.182  -2.84 0.0168   0.549  0.0345 Gaussian c(0, 1)
#>  8 1              8  0.266  -2.68 0.0233   0.557  0.0757 Gaussian c(0, 1)
#>  9 1              9 -0.255  -2.52 0.0293   0.565 -0.178  Gaussian c(0, 1)
#> 10 1             10  0.0722 -2.36 0.0339   0.573 -0.0187 Gaussian c(0, 1)
#> # … with 190 more rows