Combine Multiple Tidy Distributions of Different Types

This allows a user to specify any n number of tidy_ distributions that can be combined into a single tibble. This is the preferred method for combining multiple distributions of different types, for example a Gaussian distribution and a Beta distribution.

This generates a single tibble with an added column of dist_type that will give the distribution family name and its associated parameters.

tidy_combine_distributions(...)

Arguments

...

The ... is where you can place your different distributions

Value

A tibble

Details

Allows a user to generate a tibble of different tidy_ distributions

See also

Author

Steven P. Sanderson II, MPH

Examples

tn <- tidy_normal()
tb <- tidy_beta()
tc <- tidy_cauchy()

tidy_combine_distributions(tn, tb, tc)
#> # A tibble: 150 × 8
#>    sim_number     x      y    dx       dy     p          q dist_type       
#>    <fct>      <int>  <dbl> <dbl>    <dbl> <dbl>      <dbl> <fct>           
#>  1 1              1 -0.383 -4.00 0.000299 0.5      0.00233 Gaussian c(0, 1)
#>  2 1              2  0.566 -3.85 0.000809 0.508    0.551   Gaussian c(0, 1)
#>  3 1              3  0.104 -3.70 0.00197  0.516    0.274   Gaussian c(0, 1)
#>  4 1              4  0.403 -3.56 0.00436  0.524    0.449   Gaussian c(0, 1)
#>  5 1              5 -2.67  -3.41 0.00869  0.533 -Inf       Gaussian c(0, 1)
#>  6 1              6 -1.42  -3.26 0.0157   0.541   -0.599   Gaussian c(0, 1)
#>  7 1              7  0.521 -3.11 0.0259   0.549    0.522   Gaussian c(0, 1)
#>  8 1              8  0.324 -2.97 0.0389   0.557    0.402   Gaussian c(0, 1)
#>  9 1              9 -0.611 -2.82 0.0537   0.565   -0.123   Gaussian c(0, 1)
#> 10 1             10  0.528 -2.67 0.0685   0.573    0.527   Gaussian c(0, 1)
#> # … with 140 more rows

## OR

tidy_combine_distributions(
  tidy_normal(),
  tidy_beta(),
  tidy_cauchy(),
  tidy_logistic()
)
#> # A tibble: 200 × 8
#>    sim_number     x      y    dx       dy     p       q dist_type       
#>    <fct>      <int>  <dbl> <dbl>    <dbl> <dbl>   <dbl> <fct>           
#>  1 1              1 -0.268 -3.33 0.000251 0.5   -0.0338 Gaussian c(0, 1)
#>  2 1              2  0.142 -3.20 0.000667 0.508  0.226  Gaussian c(0, 1)
#>  3 1              3 -0.807 -3.07 0.00159  0.516 -0.382  Gaussian c(0, 1)
#>  4 1              4  0.636 -2.95 0.00342  0.524  0.563  Gaussian c(0, 1)
#>  5 1              5  0.474 -2.82 0.00663  0.533  0.447  Gaussian c(0, 1)
#>  6 1              6  1.18  -2.69 0.0117   0.541  1.04   Gaussian c(0, 1)
#>  7 1              7  0.204 -2.56 0.0187   0.549  0.266  Gaussian c(0, 1)
#>  8 1              8 -0.859 -2.44 0.0276   0.557 -0.418  Gaussian c(0, 1)
#>  9 1              9 -0.414 -2.31 0.0377   0.565 -0.126  Gaussian c(0, 1)
#> 10 1             10  1.12  -2.18 0.0484   0.573  0.973  Gaussian c(0, 1)
#> # … with 190 more rows