Compare Empirical Data to Distributions

Compare some empirical data set against different distributions to help find the distribution that could be the best fit.

tidy_distribution_comparison(.x, .distribution_type = "continuous")

Arguments

.x

The data set being passed to the function

.distribution_type

What kind of data is it, can be one of continuous or discrete

Value

An invisible list object. A tibble is printed.

Details

The purpose of this function is to take some data set provided and to try to find a distribution that may fit the best. A parameter of .distribution_type must be set to either continuous or discrete in order for this the function to try the appropriate types of distributions.

The following distributions are used:

Continuous:

tidy_beta

tidy_cauchy

tidy_exponential

tidy_gamma

tidy_logistic

tidy_lognormal

tidy_pareto

tidy_uniform

tidy_weibull

Discrete:

tidy_binomial

tidy_geometric

tidy_hypergeometric

tidy_poisson

Author

Steven P. Sanderson II, MPH

Examples

xc <- mtcars$mpg
tidy_distribution_comparison(xc, "continuous")
#> For the beta distribution, its mean 'mu' should be 0 < mu < 1. The data will
#> therefore be scaled to enforce this.
#> # A tibble: 9 × 2
#>   dist_with_params        abs_tot_deviance
#>   <chr>                              <dbl>
#> 1 Gamma c(11.47, 1.75)               0.762
#> 2 Weibull c(3.58, 22.29)             0.888
#> 3 Uniform c(31.84, 8.34)             1.37 
#> 4 Beta c(1.11, 1.58, 0)              1.65 
#> 5 Logistic c(20.09, 3.27)            2.14 
#> 6 Lognormal c(2.96, 0.29)            2.42 
#> 7 Cauchy c(19.2, 7.38)               3.34 
#> 8 Pareto c(10.4, 1.62)               3.45 
#> 9 Exponential c(0.05)                5.50 

xd <- trunc(xc)
tidy_distribution_comparison(xd, "discrete")
#> # A tibble: 4 × 2
#>   dist_with_params             abs_tot_deviance
#>   <chr>                                   <dbl>
#> 1 Poisson c(19.69)                         1.14
#> 2 Hypergeometric c(21, 11, 21)             3.92
#> 3 Binomial c(32, 0.03)                     4.48
#> 4 Geometric c(0.05)                        5.34