Tidy Distribution Summary Statistics Tibble

This function returns a summary statistics tibble. It will use the y column from the tidy_ distribution function.

Usage

tidy_distribution_summary_tbl(.data, ...)

Arguments

.data

The data that is going to be passed from a a tidy_ distribution function.

...

This is the grouping variable that gets passed to dplyr::group_by() and dplyr::select().

Value

A summary stats tibble

Details

This function takes in a tidy_ distribution table and will return a tibble of the following information:

sim_number

mean_val

median_val

std_val

min_val

max_val

skewness

kurtosis

range

iqr

variance

ci_hi

ci_lo

The kurtosis and skewness come from the package healthyR.ai

Author

Steven P. Sanderson II, MPH

Examples

library(dplyr)
#> Warning: package 'dplyr' was built under R version 4.2.3
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

tn <- tidy_normal(.num_sims = 5)
tb <- tidy_beta(.num_sims = 5)

tidy_distribution_summary_tbl(tn)
#> # A tibble: 1 × 12
#>   mean_val median_val std_val min_val max_val skewness kurtosis range   iqr
#>      <dbl>      <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl> <dbl> <dbl>
#> 1  -0.0369    -0.0298   0.966   -2.94    2.64   -0.133     2.79  5.59  1.34
#> # ℹ 3 more variables: variance <dbl>, ci_low <dbl>, ci_high <dbl>
tidy_distribution_summary_tbl(tn, sim_number)
#> # A tibble: 5 × 13
#>   sim_number mean_val median_val std_val min_val max_val skewness kurtosis range
#>   <fct>         <dbl>      <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl> <dbl>
#> 1 1           0.00357     0.0321   1.04    -2.94    2.64   -0.189     3.23  5.59
#> 2 2          -0.0175     -0.0425   0.950   -2.06    1.90   -0.156     2.52  3.96
#> 3 3          -0.0866     -0.0419   0.977   -2.09    2.30    0.178     2.62  4.39
#> 4 4          -0.00155     0.0746   0.777   -1.72    1.46   -0.186     2.57  3.19
#> 5 5          -0.0826     -0.0463   1.09    -2.66    1.98   -0.223     2.39  4.65
#> # ℹ 4 more variables: iqr <dbl>, variance <dbl>, ci_low <dbl>, ci_high <dbl>

data_tbl <- tidy_combine_distributions(tn, tb)

tidy_distribution_summary_tbl(data_tbl)
#> # A tibble: 1 × 12
#>   mean_val median_val std_val min_val max_val skewness kurtosis range   iqr
#>      <dbl>      <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl> <dbl> <dbl>
#> 1    0.223      0.354   0.756   -2.94    2.64   -0.911     4.56  5.59 0.716
#> # ℹ 3 more variables: variance <dbl>, ci_low <dbl>, ci_high <dbl>
tidy_distribution_summary_tbl(data_tbl, dist_type)
#> # A tibble: 2 × 13
#>   dist_type mean_val median_val std_val  min_val max_val skewness kurtosis range
#>   <fct>        <dbl>      <dbl>   <dbl>    <dbl>   <dbl>    <dbl>    <dbl> <dbl>
#> 1 Gaussian…  -0.0369    -0.0298   0.966 -2.94      2.64   -0.133      2.79 5.59 
#> 2 Beta c(1…   0.483      0.485    0.275  0.00509   0.997   0.0142     1.86 0.992
#> # ℹ 4 more variables: iqr <dbl>, variance <dbl>, ci_low <dbl>, ci_high <dbl>