Tidy Distribution Summary Statistics Tibble

This function returns a summary statistics tibble. It will use the y column from the tidy_ distribution function.

Usage

tidy_distribution_summary_tbl(.data, ...)

Arguments

.data

The data that is going to be passed from a a tidy_ distribution function.

...

This is the grouping variable that gets passed to dplyr::group_by() and dplyr::select().

Value

A summary stats tibble

Details

This function takes in a tidy_ distribution table and will return a tibble of the following information:

sim_number

mean_val

median_val

std_val

min_val

max_val

skewness

kurtosis

range

iqr

variance

ci_hi

ci_lo

The kurtosis and skewness come from the package healthyR.ai

Author

Steven P. Sanderson II, MPH

Examples

library(dplyr)
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

tn <- tidy_normal(.num_sims = 5)
tb <- tidy_beta(.num_sims = 5)

tidy_distribution_summary_tbl(tn)
#> # A tibble: 1 × 12
#>   mean_val median_…¹ std_val min_val max_val skewn…² kurto…³ range   iqr varia…⁴
#>      <dbl>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>   <dbl>
#> 1   -0.118   -0.0700    1.05   -3.19    2.96  -0.166    2.92  6.15  1.42    1.11
#> # … with 2 more variables: ci_low <dbl>, ci_high <dbl>, and abbreviated
#> #   variable names ¹​median_val, ²​skewness, ³​kurtosis, ⁴​variance
tidy_distribution_summary_tbl(tn, sim_number)
#> # A tibble: 5 × 13
#>   sim_number mean_…¹ media…² std_val min_val max_val skewn…³ kurto…⁴ range   iqr
#>   <fct>        <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>
#> 1 1          -0.0292  0.123    1.10    -2.69    1.72  -0.451    2.45  4.41  1.39
#> 2 2           0.139   0.167    0.895   -2.27    2.64   0.176    3.71  4.91  1.06
#> 3 3          -0.0847 -0.0171   0.970   -2.50    1.98  -0.240    2.54  4.48  1.36
#> 4 4          -0.238  -0.196    1.16    -2.15    2.96   0.371    2.72  5.11  1.85
#> 5 5          -0.379  -0.179    1.08    -3.19    1.96  -0.397    3.08  5.15  1.10
#> # … with 3 more variables: variance <dbl>, ci_low <dbl>, ci_high <dbl>, and
#> #   abbreviated variable names ¹​mean_val, ²​median_val, ³​skewness, ⁴​kurtosis

data_tbl <- tidy_combine_distributions(tn, tb)

tidy_distribution_summary_tbl(data_tbl)
#> # A tibble: 1 × 12
#>   mean_val median_…¹ std_val min_val max_val skewn…² kurto…³ range   iqr varia…⁴
#>      <dbl>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>   <dbl>
#> 1    0.205     0.365   0.835   -3.19    2.96   -1.02    4.72  6.15 0.808   0.698
#> # … with 2 more variables: ci_low <dbl>, ci_high <dbl>, and abbreviated
#> #   variable names ¹​median_val, ²​skewness, ³​kurtosis, ⁴​variance
tidy_distribution_summary_tbl(data_tbl, dist_type)
#> # A tibble: 2 × 13
#>   dist_type mean_…¹ media…² std_val  min_val max_val skewn…³ kurto…⁴ range   iqr
#>   <fct>       <dbl>   <dbl>   <dbl>    <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>
#> 1 Gaussian…  -0.118 -0.0700   1.05  -3.19e+0    2.96 -0.166     2.92 6.15  1.42 
#> 2 Beta c(1…   0.528  0.539    0.283  6.47e-4    1.00 -0.0849    1.90 0.999 0.455
#> # … with 3 more variables: variance <dbl>, ci_low <dbl>, ci_high <dbl>, and
#> #   abbreviated variable names ¹​mean_val, ²​median_val, ³​skewness, ⁴​kurtosis