Tidy Distribution Summary Statistics Tibble

This function returns a summary statistics tibble. It will use the y column from the tidy_ distribution function.

tidy_distribution_summary_tbl(.data, ...)

Arguments

.data

The data that is going to be passed from a a tidy_ distribution function.

...

This is the grouping variable that gets passed to dplyr::group_by() and dplyr::select().

Value

A summary stats tibble

Details

This function takes in a tidy_ distribution table and will return a tibble of the following information:

sim_number

mean_val

median_val

std_val

min_val

max_val

skewness

kurtosis

range

iqr

variance

The kurtosis and skewness come from the package healthyR.ai

Author

Steven P. Sanderson II, MPH

Examples

library(dplyr)
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

tn <- tidy_normal(.num_sims = 5)
tb <- tidy_beta(.num_sims = 5)

tidy_distribution_summary_tbl(tn)
#> # A tibble: 1 x 10
#>   mean_val median_val std_val min_val max_val skewness kurtosis range   iqr
#>      <dbl>      <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl> <dbl> <dbl>
#> 1  -0.0219    -0.0200    1.03   -3.30    3.57    0.104     3.33  6.86  1.38
#> # ... with 1 more variable: variance <dbl>
tidy_distribution_summary_tbl(tn, sim_number)
#> # A tibble: 5 x 11
#>   sim_number mean_val median_val std_val min_val max_val skewness kurtosis range
#>   <fct>         <dbl>      <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl> <dbl>
#> 1 1          -0.0561     -0.0321   1.20    -3.30    2.38   -0.223     3.12  5.68
#> 2 2           0.00335     0.0283   0.925   -1.76    3.57    0.977     5.86  5.32
#> 3 3           0.177       0.101    0.946   -1.69    2.74    0.135     2.83  4.43
#> 4 4          -0.201      -0.314    0.991   -2.52    1.46   -0.193     2.10  3.98
#> 5 5          -0.0327     -0.0504   1.05    -2.19    2.59    0.326     2.65  4.78
#> # ... with 2 more variables: iqr <dbl>, variance <dbl>

data_tbl <- tidy_combine_distributions(tn, tb)

tidy_distribution_summary_tbl(data_tbl)
#> # A tibble: 1 x 10
#>   mean_val median_val std_val min_val max_val skewness kurtosis range   iqr
#>      <dbl>      <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl> <dbl> <dbl>
#> 1    0.235      0.352   0.797   -3.30    3.57   -0.625     5.02  6.86 0.754
#> # ... with 1 more variable: variance <dbl>
tidy_distribution_summary_tbl(data_tbl, dist_type)
#> # A tibble: 2 x 11
#>   dist_type mean_val median_val std_val  min_val max_val skewness kurtosis range
#>   <fct>        <dbl>      <dbl>   <dbl>    <dbl>   <dbl>    <dbl>    <dbl> <dbl>
#> 1 Gaussian~  -0.0219    -0.0200   1.03  -3.30      3.57    0.104      3.33 6.86 
#> 2 Beta c(1~   0.491      0.481    0.293  0.00961   0.996   0.0175     1.72 0.987
#> # ... with 2 more variables: iqr <dbl>, variance <dbl>