Tidy Distribution Summary Statistics Tibble

This function returns a summary statistics tibble. It will use the y column from the tidy_ distribution function.

tidy_distribution_summary_tbl(.data)

Arguments

.data

The data that is going to be passed from a a tidy_ distribution function.

Value

A summary stats tibble

Details

This function takes in a tidy_ distribution table and will return a tibble of the following information:

sim_number

mean_val

median_val

std_val

min_val

max_val

skewness

kurtosis

range

iqr

variance

The kurtosis and skewness come from the package healthyR.ai

Author

Steven P. Sanderson II, MPH

Examples

tn <- tidy_normal(.num_sims = 5)
tidy_distribution_summary_tbl(tn)
#> Registered S3 method overwritten by 'tune':
#>   method                   from   
#>   required_pkgs.model_spec parsnip
#> # A tibble: 5 x 9
#>   sim_number mean_val median_val std_val min_val max_val range   iqr variance
#>   <fct>         <dbl>      <dbl>   <dbl>   <dbl>   <dbl> <dbl> <dbl>    <dbl>
#> 1 1            0.0110    -0.0502   0.879   -2.13    2.16  4.30  1.21    0.773
#> 2 2            0.177      0.278    0.951   -2.60    2.03  4.63  1.26    0.905
#> 3 3           -0.0541    -0.243    1.22    -3.09    2.92  6.00  1.79    1.49 
#> 4 4            0.0370    -0.0142   0.964   -2.42    2.37  4.79  1.32    0.929
#> 5 5           -0.158     -0.165    0.971   -2.02    2.65  4.68  1.02    0.942