Tidy Randomly Generated Inverse Exponential Distribution Tibble

This function will generate n random points from an inverse exponential distribution with a user provided, .rate or .scale and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

tidy_inverse_exponential(.n = 50, .rate = 1, .scale = 1/.rate, .num_sims = 1)

Arguments

.n

The number of randomly generated points you want.

.rate

An alternative way to specify the .scale

.scale

Must be strictly positive.

.num_sims

The number of randomly generated simulations you want.

Value

A tibble of randomly generated data.

Details

This function uses the underlying actuar::rinvexp(), and its underlying p, d, and q functions. For more information please see actuar::rinvexp()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_inverse_exponential()
#> # A tibble: 50 x 7
#>    sim_number     x      y    dx       dy        p     q
#>    <fct>      <int>  <dbl> <dbl>    <dbl>    <dbl> <dbl>
#>  1 1              1 26.6   -5.51 0.000779 0        0    
#>  2 1              2  3.83  -1.34 0.0723   5.24e-22 0.257
#>  3 1              3  0.820  2.83 0.0913   2.29e-11 0.313
#>  4 1              4  0.442  7.00 0.0239   8.06e- 8 0.358
#>  5 1              5  0.492 11.2  0.0120   4.79e- 6 0.399
#>  6 1              6  0.483 15.3  0.0138   5.55e- 5 0.438
#>  7 1              7  1.10  19.5  0.00786  2.84e- 4 0.476
#>  8 1              8  1.28  23.7  0.00215  9.12e- 4 0.514
#>  9 1              9  0.892 27.9  0.00416  2.19e- 3 0.552
#> 10 1             10  0.509 32.0  0.00401  4.32e- 3 0.590
#> # ... with 40 more rows