Tidy Randomly Generated Inverse Exponential Distribution Tibble

This function will generate n random points from an inverse exponential distribution with a user provided, .rate or .scale and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

Usage

tidy_inverse_exponential(
  .n = 50,
  .rate = 1,
  .scale = 1/.rate,
  .num_sims = 1,
  .return_tibble = TRUE
)

Arguments

.n

The number of randomly generated points you want.

.rate

An alternative way to specify the .scale

.scale

Must be strictly positive.

.num_sims

The number of randomly generated simulations you want.

.return_tibble

A logical value indicating whether to return the result as a tibble. Default is TRUE.

Value

A tibble of randomly generated data.

Details

This function uses the underlying actuar::rinvexp(), and its underlying p, d, and q functions. For more information please see actuar::rinvexp()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_inverse_exponential()
#> # A tibble: 50 × 7
#>    sim_number     x     y     dx       dy     p     q
#>    <fct>      <int> <dbl>  <dbl>    <dbl> <dbl> <dbl>
#>  1 1              1 2.62  -1.47  6.31e- 4 0.682 2.62 
#>  2 1              2 1.81   0.561 2.99e- 1 0.575 1.81 
#>  3 1              3 2.19   2.59  1.35e- 1 0.633 2.19 
#>  4 1              4 0.545  4.62  1.70e- 2 0.160 0.545
#>  5 1              5 1.31   6.65  1.35e- 2 0.465 1.31 
#>  6 1              6 0.523  8.67  1.42e- 2 0.148 0.523
#>  7 1              7 0.777 10.7   6.03e- 3 0.276 0.777
#>  8 1              8 1.10  12.7   2.03e- 8 0.403 1.10 
#>  9 1              9 1.62  14.8   2.25e-13 0.540 1.62 
#> 10 1             10 3.25  16.8   3.93e- 5 0.735 3.25 
#> # ℹ 40 more rows