Tidy Randomly Generated Gaussian Distribution Tibble

This function will generate n random points from a Gaussian distribution with a user provided, .mean, .sd - standard deviation and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the dnorm, pnorm and qnorm data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

tidy_normal(.n = 50, .mean = 0, .sd = 1, .num_sims = 1)

Arguments

.n

The number of randomly generated points you want.

.mean

The mean of the randomly generated data.

.sd

The standard deviation of the randomly generated data.

.num_sims

The number of randomly generated simulations you want.

Value

A tibble of randomly generated data.

Details

This function uses the underlying stats::rnorm(), stats::pnorm(), and stats::qnorm() functions to generate data from the given parameters. For more information please see stats::rnorm()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_normal()
#> # A tibble: 50 x 7
#>    sim_number     x       y    dx       dy     p        q
#>    <fct>      <int>   <dbl> <dbl>    <dbl> <dbl>    <dbl>
#>  1 1              1 -1.32   -3.36 0.000303 0.5   -Inf    
#>  2 1              2 -0.894  -3.24 0.000934 0.508   -2.05 
#>  3 1              3 -0.0182 -3.12 0.00246  0.516   -1.74 
#>  4 1              4 -0.194  -3.00 0.00556  0.524   -1.54 
#>  5 1              5 -0.914  -2.88 0.0108   0.533   -1.39 
#>  6 1              6  0.0416 -2.76 0.0182   0.541   -1.27 
#>  7 1              7 -0.645  -2.64 0.0269   0.549   -1.16 
#>  8 1              8  0.504  -2.52 0.0352   0.557   -1.07 
#>  9 1              9  0.707  -2.40 0.0417   0.565   -0.981
#> 10 1             10  0.981  -2.28 0.0455   0.573   -0.901
#> # ... with 40 more rows