Tidy Random Walk

Takes in the data from a tidy_ distribution function and applies a random walk calculation of either cum_prod or cum_sum to y.

Usage

tidy_random_walk(
  .data,
  .initial_value = 0,
  .sample = FALSE,
  .replace = FALSE,
  .value_type = "cum_prod"
)

Arguments

.data

The data that is being passed from a tidy_ distribution function.

.initial_value

The default is 0, this can be set to whatever you want.

.sample

This is a boolean value TRUE/FALSE. The default is FALSE. If set to TRUE then the y value from the tidy_ distribution function is sampled.

.replace

This is a boolean value TRUE/FALSE. The default is FALSE. If set to TRUE AND .sample is set to TRUE then the replace parameter of the sample function will be set to TRUE.

.value_type

This can take one of three different values for now. These are the following:

"cum_prod" - This will take the cumprod of y

"cum_sum" - This will take the cumsum of y

Value

An ungrouped tibble.

Details

Monte Carlo simulations were first formally designed in the 1940’s while developing nuclear weapons, and since have been heavily used in various fields to use randomness solve problems that are potentially deterministic in nature. In finance, Monte Carlo simulations can be a useful tool to give a sense of how assets with certain characteristics might behave in the future. While there are more complex and sophisticated financial forecasting methods such as ARIMA (Auto-Regressive Integrated Moving Average) and GARCH (Generalised Auto-Regressive Conditional Heteroskedasticity) which attempt to model not only the randomness but underlying macro factors such as seasonality and volatility clustering, Monte Carlo random walks work surprisingly well in illustrating market volatility as long as the results are not taken too seriously.

Author

Steven P. Sanderson II, MPH

Examples

tidy_normal(.sd = .1, .num_sims = 25) %>%
  tidy_random_walk()
#> # A tibble: 1,250 × 8
#>    sim_number     x        y     dx      dy     p        q random_walk_value
#>    <fct>      <int>    <dbl>  <dbl>   <dbl> <dbl>    <dbl>             <dbl>
#>  1 1              1  0.0110  -0.389 0.00194 0.5    0.0106            0.0110 
#>  2 1              2  0.151   -0.374 0.00495 0.581  0.0976            0.164  
#>  3 1              3  0.0593  -0.358 0.0115  0.658  0.0368            0.233  
#>  4 1              4  0.225   -0.343 0.0240  0.730  0.234             0.510  
#>  5 1              5 -0.114   -0.327 0.0458  0.793 -0.0585            0.337  
#>  6 1              6 -0.171   -0.312 0.0803  0.846 -0.0988            0.109  
#>  7 1              7 -0.114   -0.296 0.131   0.890 -0.0584           -0.0179 
#>  8 1              8  0.00830 -0.281 0.201   0.923  0.00919          -0.00977
#>  9 1              9  0.0477  -0.265 0.295   0.949  0.0303            0.0374 
#> 10 1             10  0.0606  -0.250 0.421   0.967  0.0375            0.100  
#> # … with 1,240 more rows
#> # ℹ Use `print(n = ...)` to see more rows