Tidy Randomly Generated T Distribution Tibble

This function will generate n random points from a rt distribution with a user provided, df, ncp, and number of random simulations to be produced. The function returns a tibble with the simulation number column the x column which corresponds to the n randomly generated points, the d_, p_ and q_ data points as well.

The data is returned un-grouped.

The columns that are output are:

sim_number The current simulation number.

x The current value of n for the current simulation.

y The randomly generated data point.

dx The x value from the stats::density() function.

dy The y value from the stats::density() function.

p The values from the resulting p_ function of the distribution family.

q The values from the resulting q_ function of the distribution family.

tidy_t(.n = 50, .df = 1, .ncp = 0, .num_sims = 1)

Arguments

.n

The number of randomly generated points you want.

.df

Degrees of freedom, Inf is allowed.

.ncp

Non-centrality parameter.

.num_sims

The number of randomly generated simulations you want.

Value

A tibble of randomly generated data.

Details

This function uses the underlying stats::rt(), and its underlying p, d, and q functions. For more information please see stats::rt()

See also

Author

Steven P. Sanderson II, MPH

Examples

tidy_t()
#> # A tibble: 50 × 7
#>    sim_number     x      y     dx          dy     p      q
#>    <fct>      <int>  <dbl>  <dbl>       <dbl> <dbl>  <dbl>
#>  1 1              1 -1.27  -16.2  0.000118    0.5   -0.352
#>  2 1              2 -4.47  -15.4  0.00133     0.506 -0.762
#>  3 1              3  0.122 -14.7  0.00589     0.513 -0.204
#>  4 1              4 -0.288 -13.9  0.0103      0.519 -0.247
#>  5 1              5 -0.799 -13.2  0.00703     0.526 -0.300
#>  6 1              6  0.334 -12.4  0.00190     0.532 -0.183
#>  7 1              7 -0.382 -11.7  0.000202    0.539 -0.256
#>  8 1              8  6.42  -10.9  0.00000841  0.545  0.441
#>  9 1              9 -0.721 -10.2  0.000000583 0.552 -0.292
#> 10 1             10  0.822  -9.43 0.0000212   0.558 -0.134
#> # … with 40 more rows